refactor: create abstract form input
