+++ src/app/shared/components/ui/form/mask-field/mask-field.component.html
@@ -12,7 +12,7 @@
     [autoClear]="autoClear()"
     [disabled]="disabled"
     [(ngModel)]="value"
-    (onInput)="onInputMaskChange($event)"
+    (onInput)="onInputChange($event)"
     (focus)="markAsTouched()"
   />
 </div>
