+++ src/app/shared/components/ui/form/password-field/password-field.component.html
@@ -11,7 +11,7 @@
     [toggleMask]="toggleMask()"
     [disabled]="disabled"
     formControlName="password"
-    (input)="onInputPasswordChange($event)"
+    (input)="onInputChange($event)"
     (focus)="markAsTouched()"
   />
 </div>
